requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
joblib
nltk

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
python-dotenv
spacy
googletrans==4.0.0-rc1
langdetect

# Web, API
fastapi
uvicorn
streamlit
bs4

# postgresql
psycopg2-binary
